Delete ShimCal
The class ShimCal effectively outputs a ShimOpt_ type object.  This
doesn't make sense and, in any case, ShimCal is simple enough just to be
integrated with ShimOpt_. It's likely that the only shim system-specific
classes required will be ShimOpt_XXX and ShimCom_XXX. This commit is the
1st step: Deleting ShimCal.m and ShimCal_IUGM_Prisma_fit.m (which is
incorporated into ShimOpt_IUGM_Prisma_fit.m).