Delete ShimCal - step 2
All remaining ShimCal_ objects deleted. Some of ShimCal_Acdc was
incorporated into ShimOpt_Acdc, but some of this copied code will soon
be changed again anyway (as the shim reference map data will be
reacquired in the coming days). Also deleted all of Shim_Martinos
related material. It is mostly deprecated with respect to the super
classes and can be reintroduced once the latter have a stable version.